requirements.txt
Flask==3.1.2
Flask_PyMongo==2.3.0
flask_session==0.8.0
geojson==3.1.0
geopandas==1.1.1
matplotlib==3.10.5
networkx==3.4.2
osmnx==2.0.6
pymongo==4.14.1
python_bcrypt==0.3.2
